autoregistrator moved to website (no longer separate program)
