+++ Fixer/Fixer.csproj
@@ -97,10 +97,6 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\AutoRegistrator\AutoRegistrator.csproj">
-      <Project>{a243d9f0-b7fa-4db9-ad98-ddde85624049}</Project>
-      <Name>AutoRegistrator</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Benchmarker\Benchmarker.csproj">
       <Project>{67501a5b-b6bb-467e-85da-22d81c40829c}</Project>
       <Name>Benchmarker</Name>
