+++ Zero-K.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlasmaShared", "Shared\PlasmaShared\PlasmaShared.csproj", "{B24E7FB3-A3FB-4B67-A499-721F9B34A404}"
 EndProject
@@ -26,8 +26,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GalaxyDesigner", "GalaxyDesigner\GalaxyDesigner.csproj", "{BB4AD85D-23E0-429A-8FE4-74E2A58757D5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlanetWarsReplayCreator", "PlanetWarsReplayCreator\PlanetWarsReplayCreator.csproj", "{83FBD383-4A72-417B-B77E-B35FCFFA6A56}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AutoRegistrator", "AutoRegistrator\AutoRegistrator.csproj", "{A243D9F0-B7FA-4DB9-AD98-DDDE85624049}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{90D90F08-B8EA-4777-BF68-C3DE8DC27427}"
 EndProject
@@ -98,10 +96,6 @@ Global
 		{83FBD383-4A72-417B-B77E-B35FCFFA6A56}.Debug|x86.Build.0 = Debug|x86
 		{83FBD383-4A72-417B-B77E-B35FCFFA6A56}.Release|x86.ActiveCfg = Release|x86
 		{83FBD383-4A72-417B-B77E-B35FCFFA6A56}.Release|x86.Build.0 = Release|x86
-		{A243D9F0-B7FA-4DB9-AD98-DDDE85624049}.Debug|x86.ActiveCfg = Debug|x86
-		{A243D9F0-B7FA-4DB9-AD98-DDDE85624049}.Debug|x86.Build.0 = Debug|x86
-		{A243D9F0-B7FA-4DB9-AD98-DDDE85624049}.Release|x86.ActiveCfg = Release|x86
-		{A243D9F0-B7FA-4DB9-AD98-DDDE85624049}.Release|x86.Build.0 = Release|x86
 		{90D90F08-B8EA-4777-BF68-C3DE8DC27427}.Debug|x86.ActiveCfg = Debug|x86
 		{90D90F08-B8EA-4777-BF68-C3DE8DC27427}.Debug|x86.Build.0 = Debug|x86
 		{90D90F08-B8EA-4777-BF68-C3DE8DC27427}.Release|x86.ActiveCfg = Release|x86
